+++ CHANGELOG.md
@@ -6,6 +6,7 @@ Updates should follow the [Keep a CHANGELOG](http://keepachangelog.com/) princip
 
 ## [Unreleased]
 - Add `EnvelopeDocumentAssignmentsEndpoint`
+- Add `AccountBrandings` CRUD endpoint
 
 ## [1.9.0]
 ### Added